Enable -Wall -Werror by default

+++ apps/http-socket/http-socket.c
@@ -442,12 +442,13 @@ event(struct tcp_socket *tcps, void *ptr,
         tcp_socket_send_str(tcps, "Range: bytes=");
         if(s->length) {
           if(s->pos >= 0) {
-            sprintf(str, "%llu-%llu", s->pos, s->pos + s->length - 1);
+            sprintf(str, "%llu-%llu",
+              (long long unsigned int)s->pos, (long long unsigned int)s->pos + s->length - 1);
           } else {
-            sprintf(str, "-%llu", s->length);
+            sprintf(str, "-%llu", (long long unsigned int)s->length);
           }
         } else {
-          sprintf(str, "%llu-", s->pos);
+          sprintf(str, "%llu-", (long long unsigned int)s->pos);
         }
         tcp_socket_send_str(tcps, str);
         tcp_socket_send_str(tcps, "\r\n");
