+++ core/net/mac/tsch/tsch-adaptive-timesync.c
@@ -101,7 +101,7 @@ timesync_learn_drift_ticks(uint32_t time_delta_asn, int32_t drift_ticks)
 
   TSCH_LOG_ADD(tsch_log_message,
       snprintf(log->message, sizeof(log->message),
-          "drift %ld", drift_ppm / 256));
+          "drift %ld", (long int)drift_ppm / 256));
 }
 /*---------------------------------------------------------------------------*/
 /* Either reset or update the neighbor's drift */
@@ -147,7 +147,7 @@ compensate_internal(uint32_t time_delta_usec, int32_t drift_ppm, int32_t *remain
   if(ABS(amount_ticks) > RTIMER_ARCH_SECOND / 128) {
     TSCH_LOG_ADD(tsch_log_message,
         snprintf(log->message, sizeof(log->message),
-            "!too big compensation %ld delta %ld", amount_ticks, time_delta_usec));
+            "!too big compensation %ld delta %ld", (long int)amount_ticks, (long int)time_delta_usec));
     amount_ticks = (amount_ticks > 0 ? RTIMER_ARCH_SECOND : -RTIMER_ARCH_SECOND) / 128;
   }
